Root path redirects to a successful endpoint
When booting up the app for the first time it is nice to know you've got through the set up steps correctly by checking you've got a working thing.

It is somewhat conventional to boot a Rails/Puma app and check the default URI of localhost:3000 to see if it worked **before** reading further in the documentation to see what the endpoints yoi might like to use are. Redirect to the health check as a basic sign of health to meet this need.

I noticed the Visitors controller had been removed (it exists in the Rails Template)  but this old route still existed. Whenever someone used it, it failed.