Follow Template TODO and rename application
Apart from this being the guidance for using the Rails Template, this
renaming will serve to properly seperate this app and it's dependencies
from others running on your machine. For example if 2 apps miss changing
their DATABASE_URL then both will try to create and use the same
database name - if not using Docker.