CI tests run in the TEST environment rather than DEVELOPMENT
Create a new Docker Compose file for the test environment. This enables
the test suite to be run locally and on CI with the right RAILS_ENV of
TEST. It also allows us to start diverging what environment variables
should be made available since only the since .env was being used in
both. We should avoid using real environment variables in test runs just
in case real requests are performed.