Handle mismatched lists of periods per schedule
If the timescale covered by the periods in each schedule
vary e.g.

- OOH first line from Dec 1 2020 to Dec 31 2020
- OOH second line from Dec 1 2020 to Jan 31 2021

we would see a

  NoMethodError: undefined method `user' for nil:NilClass

coming out of

  lib/patterdale/out_of_hours/rotations.rb:95:in
    `find_user_for_date'

when looking for the first line user for Jan 1

our solution is to use an `UnknownUser` null object. We set an
instance of this class to the `UNKNOWN_USER` constant as the
`chunked_schedule` breaks up the rota into elements based on
unique `User` objects. i.e. whenever the composition of the
team changes, a new rota is represented.