Rename second line OOH schedule ID env var
Use a consistent name for the env var referring to the ID of the OpsGenie schedule for second line out of hours support.

This schedule includes one timeline, for the second line out of hours support. (Which is not split between dev and ops, because it is a devops kind of thing.)